fix: Refactoring
