+++ project/src/components/film-card/film-card-overview/film-card-overview.tsx
@@ -1,5 +1,5 @@
-import FilmRating from '../film-rating/film-rating';
-import {Film} from '../../types/film';
+import FilmRating from '../../film-rating/film-rating';
+import {Film} from '../../../types/film';
 
 type Props = {
   film: Film
